Increment 2: refactoring
