+++ internal/app/repositories/repositories.go
@@ -1,0 +1,6 @@
+package repositories
+
+type LinksRepository interface {
+	Put(string) string
+	Get(string) (string, bool)
+}
